add a hack to support complicated `in` and enable a test:
each object, which supports `is` should have both `__contains__`, which is a regular python method and `contains-hack`, which does the same but is just a EO method. This is needed because currently it is only possible to use an identifier to address a python method. That is, calling "sring literal".__contains__(...) does not work now because of a limitation of a pass, which adds `self` to each method call. On the other hand, calling EO method is possible in any context, like `(pystring "sring literal").contains-hack ...` is ok.